Fix/sp chars spm (#908)
* allow to put in literal strings to bypass pieces

sometimes you want to request a literal.  since
SPM has the same behavior for full string and space
delimited sub-pieces, this is relatively easy to do

* fix vec len issue

* add Pile reader

* allow GPT2 splitting on vec for SPM and fix RA arg

* force int32

* Update setup.py

listify

* dont retok

* avoid pep517 for -e install

+++ mead/api_examples/pretrain_tlm_tf.py
@@ -56,12 +56,12 @@ feature_description = {
 
 def _parse_tf_record(example_proto):
     record = tf.io.parse_single_example(example_proto, feature_description)
-    return record['x'], record['y']
+    return tf.cast(record['x'], dtype=tf.int32), tf.cast(record['y'], dtype=tf.int32)
 
 
 def _parse_tf_record_causal(example_proto):
     record = tf.io.parse_single_example(example_proto, feature_description)
-    return record['x'][:-1], record['x'][1:]
+    return tf.cast(record['x'][:-1], dtype=tf.int32), tf.cast(record['x'][1:], dtype=tf.int32)
 
 
 def decode_json(example):
